fix integer overflow and fast halko plus some optimization

+++ src/Arnoldi.hpp
@@ -11,8 +11,8 @@ public:
 
     ~ArnoldiOpData() {}
 
-    inline uint rows() const { return n; }
-    inline uint cols() const { return n; }
+    inline uint64 rows() const { return n; }
+    inline uint64 cols() const { return n; }
     // y = G * G' * x ; data.G is n x m;
     void perform_op(const float *x_in, float* y_out);
     inline void setFlags(bool is_update, bool is_standardize, bool is_pcangsd) {update = is_update; standardize = is_standardize; pcangsd = is_pcangsd;}
@@ -23,7 +23,7 @@ public:
 
 private:
     Data* data;
-    const uint n;
+    const uint64 n;
     bool update = false, standardize = false, pcangsd = false;
 };
